requirements.txt
requests
bs4
pandas
openpyxl
python-dateutil
lxml
python-dotenv